Fix data-int IP range for validation webhooks
The firewall rule has to allow connections from the Kubernetes
master to the pods in the cluster so that it can query validation
webhooks.  The IP range in the rule for integration was copied from
dev, instead of matching the master IP range for integration (which
is not the same).