Use postgres-sql for RSP database instance
postgres-private attempts to set up a private networking resource
that is already managed by the separate CloudSQL deployment, and
therefore conflicts.  Enhance postgres-sql to take the same new
parameters we're using and use it instead, and directly create the
service account (since this conceptually isn't part of the database
anyway).