Enable random instance names for RSP database
Otherwise, Terraform cannot safely delete and recreate the database
since database instance names cannot be reused for a week after they
are destroyed.  While that will be uncommon, it's very obnoxious to
not be able to do so when it becomes necessary, so we'll take the
configuration hit.

Force a Terraform run for all Science Platform environments by
removing blank lines.