Set maintenance window for data-dev GKE
We have to have 48 hours of maintenance in any 32 day interval, so
satisfy that with 12 hours per week.  Plumb the maintenance_*
settings through the module so that they can be set per environment.
data-dev will use midnight to noon UTC Wednesday (5pm to 5am PDT
Tuesday to Wednesday), and (later) int will do the same Thursday
and prod Friday.