Add GCS bucket for cutouts
This is rather unfortunately mixed in with CloudSQL because it
needs to reuse the service account created for PostgreSQL access
for GCS bucket access as well.  Object creation access is granted
to the same service account as is currently used for Butler access.
This is temporary until we have client/server Butler.