Add additional IAM binding for vo-cutouts account
We're using workload identity to bind a GKE service account to the
Google service account for vo-cutouts, which in addition to having
Cloud SQL access also has GCS access and needs to be able to
generate signed GCS URLs.

To do URL signing, it must have a private key, but exporting the
private key of a Google service account is deprecated and less secure,
and workload identity does not provide the private key.  The
alternative is to grant the service account the ability to create
tokens for itself (via the role iam.ServiceAccountTokenAccess), which
allows it to use its GKE credentials to bootstrap to a private key
that in turn can generate a pre-signed URL.

Enable this on dev and int.  Tested by manually creating the
equivalent policy on int.