Add HiPS bucket for DP0.2
Use the naming convention discussed on Slack.  Grant RW access to
the Butler service account we've been using for ease of importing
data (the plan is to remove that access once the data is imported).

Create the crawlspace service account for data-int and add it to
the HiPS ACL list.