DM-50602: Grafana Postgres DB
Grafana will be deployed in every Phalanx env.
It can use a SQLite DB, but:
* It doesn't work very well over networked storage
* It won't scale if Grafana gets a lot of usage, and we'll have to move
  to Postgres eventually anyway
* Alerting functionality works better with a Postgres instance

Eventually we want a separate database instance for Grafana so that we don't lose grafana when this database instance is under pressure (which is one situation where we will probably want it most), but let's just get it off the ground for DP1.