DM-51219: Roundtable Atlantis instance monitoring.admin in all projects
The terraform config for provisioning Google Cloud monitoring resources
is going to live in [Prodromos](https://prodromos.lsst.io), because
Prodromos is the place where we're experimenting with using Terraform to
provision monitoring resources.

The Terraform config in Prodromos gets provisioned using
[Atlantis](https://runatlantis.io), specifically the Atlantis instance
in `roundtable-prod`.

To provision Google Cloud monitoring resources via Atlantis, the Google
Cloud service account that is federated with the Atlantis pod service
account must have `monitoring.admin` powers in ALL of the projects that
we Atlantis to provision.