Resources for GKE cluster ingress static IPs
Creating a LoadBalancer Service in a GKE cluster will automatically provision an ephemeral GKE static IP address that will disappear when the Kubernetes LoadBalancer Service is destroyed. We have been manually converting these static IP addresses to Capture all the static GKE IP addresses in Terraform so that they won’t disappear if the load balancers ever accidentally get destroyed.

Let’s capture these in Terraform so that we don’t have to remember to manually make them static for any new envs we stand up, and so that they get reprovisioned automatically if we want to destroy and rebuild the entire env (though the IP address would change in that case).