Sasquatch static IPs
For Kafka and InfluxDB.

Creating a Strimzi Kafka instance in a in a GKE cluster will automatically provision a`LoadBalancer` service, which will automatically provision some ephemeral GKE static IP addresss that will disappear when the Strimzi instance is destroyed. We have been manually converting these static IP addresses in the Google Cloud web console to make sure they won’t disappear if the load balancers ever accidentally get destroyed.

Let’s capture these in Terraform so that we don’t have to remember to manually make them static for any new envs we stand up, and so that they get reprovisioned automatically if we want to destroy and rebuild the entire env (though the IP address would change in that case).