Capture enabled GKE Backup APIs in Terraform config
We have to enable the GKE Backup API in all Google projects where we
have clusters if we want to be able to rebuild them. This API seems to
be enabled by default, but it is not specified in our Terraform config.
We should specify it there to adhere to our pattern of covering all
Google Cloud changes in Terraform for disaster recovery purposes.