Add GKE Backup permissions to GitHub pipelines accounts
GitHub pipelines should be using specific service accounts for each
environment when applying Terraform config. These accounts now need
permissions on GKE Backup resources.

Some pipelines are still using the `terraform-sa` account, which should
only be used when bootstrapping environments, which is why the apply of
these GKE backup resources worked in some environments.

Note that there is no pipeline service account for `idfprod`, so that's
why there is no change in this PR for `idfprod`.