Ignore changes to cluster in GKE Backup Plan resource
If you destroy the associated cluster, terraform will try to destroy and
recreate this backup plan, which will also try to destroy all of the
backups associated with the plan. If we are trying to intentionally
rebuild a cluster, we will need to destroy it first, and we don't want
this backup plan destroyed.